+++ fe/package.json
@@ -61,5 +61,5 @@
     "eslint-plugin-react": "^7.32.2",
     "prettier": "^2.8.8"
   },
-  "proxy": "http://ec2-15-165-108-106.ap-northeast-2.compute.amazonaws.com:8080/"
+  "proxy": "http://ec2-13-125-229-143.ap-northeast-2.compute.amazonaws.com:8080/"
 }
